requirements.txt
apache-airflow-providers-apache-spark==4.1.0
pyspark==3.5.3
boto3
pandas
minio